requirements.txt
numpy
pyyaml==5.1
h5py
tensorboardX
opencv-contrib-python==4.5.2.52
tqdm